complete cabin

+++ the-wilde-oasis/src/services/apiCabins.js
@@ -12,30 +12,22 @@ export async function getCabins() {
 }
 
 export async function createEditCabin(newCabin, id) {
-  console.log(newCabin, id);
-
   const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
-
-  const imageName = newCabin.image`${Math.random()}-${
-    newCabin.image.name
-  }`.replaceAll('/', '');
-
-  // https://zrsjfukvwepvkxqccbyc.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg?t=2023-08-19T13%3A47%3A17.527Z
-
+  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
+    '/',
+    ''
+  );
   const imagePath = hasImagePath
-    ? newCabin?.image
+    ? newCabin.image
     : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  // create / edit cabin
-
+  // 1. Create/edit cabin
   let query = supabase.from('cabins');
 
-  //   A) Create Cabin
-
+  // A) CREATE
   if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
 
-  // B) edit Cabin
-
+  // B) EDIT
   if (id) query = query.update({ ...newCabin, image: imagePath }).eq('id', id);
 
   const { data, error } = await query.select().single();
@@ -45,17 +37,22 @@ export async function createEditCabin(newCabin, id) {
     throw new Error('Cabin could not be created');
   }
 
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from('cabin-images')
     .upload(imageName, newCabin.image);
 
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from('cabins').delete().eq('id', data.id);
-
+    console.error(storageError);
     throw new Error(
-      'Cabin image could not be uploaded and the the cabin was not created'
+      'Cabin image could not be uploaded and the cabin was not created'
     );
   }
+  return data;
 }
 
 export async function deleteCabin(id) {
